Tools/scripts/AP_AerobridgeTrustedFlight/requirements.txt
mavsdk==0.22.3
requests==2.27.1
python-dotenv==0.19.2
pyjwt==2.3.0
cryptography==36.0.1
arrow==1.2.2
